philo2 working

+++ philo_two/srcs/lib/ft_putnbr_fd.c
@@ -6,23 +6,23 @@
 /*   By: frthierr <frthierr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/24 14:37:15 by frthierr          #+#    #+#             */
-/*   Updated: 2020/09/09 16:21:14 by frthierr         ###   ########.fr       */
+/*   Updated: 2020/09/16 13:50:22 by frthierr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philo_one.h"
+#include "philo_two.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long n, int fd)
 {
-	unsigned int	nbr;
+	unsigned long	nbr;
 
 	if (n < 0)
 	{
 		ft_putchar_fd('-', fd);
-		nbr = (unsigned int)(n * -1);
+		nbr = (unsigned long)(n * -1);
 	}
 	else
-		nbr = (unsigned int)n;
+		nbr = (unsigned long)n;
 	if (nbr >= 10)
 		ft_putnbr_fd(nbr / 10, fd);
 	ft_putchar_fd((char)(nbr % 10 + '0'), fd);
